Fix package naming error.
