--- framework/modules/ossutils-starter/src/main/java/groop/idealworld/dew/ossutils/general/OssClientInitProcess.java
@@ -1,16 +1,0 @@
-package groop.idealworld.dew.ossutils.general;
-
-import groop.idealworld.dew.ossutils.config.OssConfigProperties;
-
-/**
- * @author yiye
- */
-public interface OssClientInitProcess {
-    /**
-     * 初始化原始客户端
-     *
-     * @param config 配置
-     * @return 是否初始化成功
-     */
-    boolean initClient(OssConfigProperties config);
-}
