--- framework/modules/ossutils-starter/src/main/java/groop/idealworld/dew/ossutils/handle/DewOssHandleClient.java
@@ -1,19 +1,0 @@
-package groop.idealworld.dew.ossutils.handle;
-
-
-/**
- * @param <T> 客户端类型
- * @author yiye
- **/
-public class DewOssHandleClient<T> {
-    private T ossClient;
-
-    public T getOssClient() {
-        return this.ossClient;
-    }
-
-    public void setOssClient(T ossClient) {
-        this.ossClient = ossClient;
-    }
-
-}
