+++ framework/modules/ossutils-starter/src/main/java/group/idealworld/dew/ossutils/handle/DewOssHandleClient.java
@@ -1,0 +1,19 @@
+package group.idealworld.dew.ossutils.handle;
+
+
+/**
+ * @param <T> 客户端类型
+ * @author yiye
+ **/
+public class DewOssHandleClient<T> {
+    private T ossClient;
+
+    public T getOssClient() {
+        return this.ossClient;
+    }
+
+    public void setOssClient(T ossClient) {
+        this.ossClient = ossClient;
+    }
+
+}
